+++ src/assisted_operator/matter/attribute.py
@@ -2,7 +2,7 @@ import numpy as np
 from src.data import Matter
 
 
-def set_color_add(m: Matter, color_add: np.int8) -> Matter:
+def set_color_add(m: Matter, color_add: int) -> Matter:
     new_matter: Matter
     new_matter = m.copy()
     new_matter.color_add = color_add
