+++ src/operator/case/reduce/bitwise.py
@@ -9,7 +9,7 @@ def bitwise_or(c: Case) -> Case:
     m0: Matter
     m1: Matter
     m0, m1 = c.matter_list[:2]
-    new_value = array_reduce.bitwise_or(m0.bool_represents(), m1.bool_represents()).astype(np.int8) * c.color_add
+    new_value = array_reduce.bitwise_or(m0.bool_represents(), m1.bool_represents()).astype(np.int) * c.color_add
     new_case = Case()
     new_case.initialize(new_value, c.background_color)
     return new_case
@@ -20,7 +20,7 @@ def bitwise_and(c: Case) -> Case:
     m0: Matter
     m1: Matter
     m0, m1 = c.matter_list[:2]
-    new_value = array_reduce.bitwise_and(m0.bool_represents(), m1.bool_represents()).astype(np.int8) * c.color_add
+    new_value = array_reduce.bitwise_and(m0.bool_represents(), m1.bool_represents()).astype(np.int) * c.color_add
     new_case = Case()
     new_case.initialize(new_value, c.background_color)
     return new_case
@@ -31,7 +31,7 @@ def bitwise_xor(c: Case) -> Case:
     m0: Matter
     m1: Matter
     m0, m1 = c.matter_list[:2]
-    new_value = array_reduce.bitwise_xor(m0.bool_represents(), m1.bool_represents()).astype(np.int8) * c.color_add
+    new_value = array_reduce.bitwise_xor(m0.bool_represents(), m1.bool_represents()).astype(np.int) * c.color_add
     new_case = Case()
     new_case.initialize(new_value, c.background_color)
     return new_case
@@ -42,7 +42,7 @@ def bitwise_not_xor(c: Case) -> Case:
     m0: Matter
     m1: Matter
     m0, m1 = c.matter_list[:2]
-    new_value = array_reduce.bitwise_not_xor(m0.bool_represents(), m1.bool_represents()).astype(np.int8) * c.color_add
+    new_value = array_reduce.bitwise_not_xor(m0.bool_represents(), m1.bool_represents()).astype(np.int) * c.color_add
     new_case = Case()
     new_case.initialize(new_value, c.background_color)
     return new_case
@@ -53,7 +53,7 @@ def bitwise_not_or(c: Case) -> Case:
     m0: Matter
     m1: Matter
     m0, m1 = c.matter_list[:2]
-    new_value = array_reduce.bitwise_not_or(m0.bool_represents(), m1.bool_represents()).astype(np.int8) * c.color_add
+    new_value = array_reduce.bitwise_not_or(m0.bool_represents(), m1.bool_represents()).astype(np.int) * c.color_add
     new_case = Case()
     new_case.initialize(new_value, c.background_color)
     return new_case
@@ -64,7 +64,7 @@ def bitwise_not_and(c: Case) -> Case:
     m0: Matter
     m1: Matter
     m0, m1 = c.matter_list[:2]
-    new_value = array_reduce.bitwise_not_and(m0.bool_represents(), m1.bool_represents()).astype(np.int8) * c.color_add
+    new_value = array_reduce.bitwise_not_and(m0.bool_represents(), m1.bool_represents()).astype(np.int) * c.color_add
     new_case = Case()
     new_case.initialize(new_value, c.background_color)
     return new_case
